Restructured frontend, minor FE/BE fixes
