+++ frontend/src/components/ListView.js
@@ -1,4 +1,4 @@
-import React, { Component } from "react";
+import React from "react";
 import Home from "./Home";
 import ReactDOM from "react-dom";
