+++ frontend/src/index.js
@@ -1,7 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import Home from "./Home";
-// import LocationForm from "./LocationForm";
+import Home from "./components/Home";
 import "./index.css";
 import "leaflet/dist/leaflet.css";
